requirements.txt
Django==4.2.29
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
